move everything to uar folder
