+++ data/uar/quiz/quiz-double-flagship-design-illustration.yml
