+++ data/uar/quiz/quiz-double-flagship-photo-illustration.yml
