+++ data/uar/quiz/quiz-single-flagship-photo.yml
