+++ data/uar/quiz/quiz-triple-flagship-photo-video-pdf.yml
