+++ features/uar/result.spec.js
@@ -6,7 +6,7 @@ module.exports = {
       name: '@result single template',
       path: '/creativecloud/quiz-recommender/result.html#type=',
       tags: '@result-single-template @result-static',
-      data: 'data/cc/quiz/result-single-template.json',
+      data: 'data/uar/quiz/result-single-template.json',
       envs: '@adobe_prod @adobe_stage',
     },
   ],
